pyproject.toml
[project]
name = "papermc-plugin-manager"
version = "0.1.0"
description = "A modern CLI tool for managing PaperMC server plugins"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "logzero>=1.7.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "typer>=0.20.0",
]

[project.scripts]
ppm = "papermc_plugin_manager.__main__:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.8.4",
]

[tool.ruff]
# Set the target Python version
target-version = "py313"

# Line length
line-length = 120

# Enable specific rule sets
lint.select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "RET",  # flake8-return
    "PTH",  # flake8-use-pathlib
]

# Ignore specific rules that may be too strict
lint.ignore = [
    "E501",    # line too long (handled by formatter)
    "B008",    # do not perform function calls in argument defaults
    "PTH123",  # open() should be replaced by Path.open()
    "B019",    # lru_cache on methods can lead to memory leaks (acceptable for connectors)
    "B904",    # raise ... from err or from None (stylistic choice)
    "N802",    # function name should be lowercase (HEADERS is conventional for HTTP headers)
    "N818",    # exception names should end with Error (PPMException is clear enough)
    "F403",    # star imports (used intentionally in __init__)
    "PTH",     # pathlib conversions (will be done gradually)
]

# Exclude directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.egg-info",
    ".pytest_cache",
]

# Format configuration
format.quote-style = "double"
format.indent-style = "space"
